requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
PyPDF2
requests
streamlit
python-multipart
huggingface_hub
langchain-community
python-dotenv
PyMuPDF
